+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-## 6.2.1
+## 1.0.0
 
-* IProov Biometrics SDK Android 6.2.1
+* IProov Biometrics SDK Android 6.3.1
+* IProov Biometrics SDK iOS 8.2.0